refactor: extract Visibility & Map Service from GameManager
Extract visibility and map-related operations into VisibilityMapService to
continue GameManager decomposition following the established refactoring plan.

Changes:
- Create VisibilityMapService with extracted visibility and map operations
  - getPlayerMapView(), getTileVisibility(), updatePlayerVisibility()
  - getMapData(), getPlayerVisibleTiles()
- Update GameManager to delegate to VisibilityMapService
- Register service in ServiceRegistry for proper lifecycle management
- Maintain all existing functionality for player visibility and map access

Progress:
- GameManager reduced from 1,557 to 1,500 lines (57 line reduction)
- Total progress: 564 lines extracted from original 2,064 lines (27% reduction)
